+++ client-app/shared/layout/components/header/_internal/mobile-menu/mobile-menu-link.vue
